backend/requirement.txt
asgiref==3.8.1
Django==5.0.7
django-autoslug==1.9.9
django-cors-headers==4.4.0
django-jazzmin==3.0.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
pillow==10.4.0
PyJWT==2.8.0
pytz==2024.1
sqlparse==0.5.1
typing_extensions==4.9.0
tzdata==2024.1
